requirements.txt
newspaper3k==0.2.8
PyPDF2==3.0.1
streamlit==1.39.0
transformers==4.45.2
validators==0.34.0
